refactor: Reorganize project directory structure for better modularity
- Create architecture-specific directories under src/arch/x86/
  - boot/: Bootloader and startup code
  - cpu/: CPU-specific features (GDT, IDT, TSS)
  - asm/: Assembly routines for low-level operations
- Reorganize kernel subsystems into logical groups:
  - core/: Main kernel entry point
  - memory/: Memory management (PMM, VMM, heap)
  - process/: Process management (scheduler, syscalls)
  - interrupt/: Interrupt handling (PIC, timer)
  - io/: I/O subsystems (terminal, keyboard)
- Restructure include directory to match source organization:
  - arch/x86/: Architecture-specific headers
  - kernel/: Kernel subsystem headers
  - lib/: Library headers
- Update Makefile to support new directory structure
- Update all include paths to reflect new organization

This refactoring improves code organization, makes it easier to add
new architectures in the future, and groups related functionality
together for better maintainability.
