+++ src/include/lib/string.h
